# Conflicts:
#	.idea/vcs.xml

+++ app/src/main/java/com/example/wildthingapp/MainActivity.java
@@ -95,9 +95,9 @@ public class MainActivity extends AppCompatActivity {
 
     }
 
-    private void sendData(float[] a){
-        float num1 = 0;
-        float num2 = 1;
+    private void sendData(double[] a){
+        double num1 = 0;
+        double num2 = 1;
 
         String num1String = a[0]*100 + "%";
         String num2String = a[1]*100 + "%";
@@ -145,7 +145,7 @@ public class MainActivity extends AppCompatActivity {
             }
 
             // Process the current movement sample in the batch (position -1)
-            float[] data = controller.processJoystickInput(event, -1);
+            double[] data = controller.processJoystickInput(event, -1);
             sendData(data);
 
             return true;
